ONL-4029 fix: Not modify the existing DEFAULT properties
