+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2020-01-16
+### Fixes
+- Fix to not modify the existing DEFAULT properties.
+
 ## [2.0.0] - 2020-01-03
 ### Fixes
 - Upgrade vulnerable versions found on dependencies.
